Feature: AutoFlow
Allows to assign the orientation/direction of the section animation during overlapping at the session level, promoting greater dynamism and customization.

A class must be added to the sf-section element to reference the animation origin. These are:
sf-autoflow-left: from left to right;
sf-autoflow-right: from right to left;
sf-autoflow-top: from top to bottom;
sf-autoflow-bottom: from bottom to top;

If no class is assigned to the section, horizontal control, either via the scrollflow element class or upon initialization in JavaScript, follows the control.